requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
google-generativeai
PyMySQL
cryptography
python-dotenv